+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 ## mini.surround
 
 - BREAKING: always move cursor to the right of left surrounding in `add()`, `delete()`, and `replace()` (instead of moving only if it was on the same line as left surrounding).
+- BREAKING: deprecate `config.funname_pattern` option in favor of manually modifying `f` surrounding.
 - Update process of getting user input: allow `<C-c>` to cancel and make empty string a valid input.